break rust code into separate modules, bump dependencies
