+++ Cargo.toml
@@ -9,9 +9,8 @@ name = "attoworld"
 crate-type = ["cdylib"]
 
 [dependencies]
-numpy = "0.25.0"
-pyo3 = "0.25.1"
-rand = "0.9.2"
-rayon = "1.11.0"
-rustfft = { version = "6.4.0", features = ["wasm_simd"] }
-
+numpy = "0.28.0"
+pyo3 = "0.28.3"
+rand = "0.10.1"
+rayon = "1.12.0"
+rustfft = { version = "6.4.1", features = ["wasm_simd"] }
